[REFACTOR]: Separate UseDropdown hook
