+++ src/types/ui.ts
@@ -43,3 +43,12 @@ export interface ColorProps {
 }
 
 export type Dispatcher<S> = React.Dispatch<React.SetStateAction<S>>;
+
+export interface DropDownContent {
+  id: string
+  label: string
+}
+
+export interface DropDownProps {
+  items: DropDownContent[]
+}
